BIG refactoring, all "id"s changed to teamId, cypherId, ...reseted relations between tables, to worked in comment from previons pull reqs, changed, removed and not reviewed tests

+++ src/main/java/dk/cngroup/lentils/repository/CypherRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CypherRepository extends JpaRepository<Cypher, Integer> {
+public interface CypherRepository extends JpaRepository<Cypher, Long> {
     public Cypher findByStage(Integer stage);
+
 }
